initialisation de player

+++ model/src/main/java/up/TowerDefense/model/object/Wall.java
@@ -3,7 +3,7 @@ package up.TowerDefense.model.object;
 public class Wall extends PlaceableObstacle{
 
     private static int newHealthScale = 5;
-    private static int newBuyingCost = 100;
+    private static double newBuyingCost = 100;
     private static int[] newSize = new int []{64,64};
 
     public Wall(double x, double y) {
